+++ front/lib/api/assistant/conversation/title.ts
@@ -9,16 +9,16 @@ import { ConversationResource } from "@app/lib/resources/conversation_resource";
 import logger from "@app/logger/logger";
 import type { ConversationType, Result } from "@app/types";
 import { Err, getLargeNonAnthropicWhitelistedModel, Ok } from "@app/types";
-import type { RunAgentArgs } from "@app/types/assistant/agent_run";
+import type { AgentLoopArgs } from "@app/types/assistant/agent_run";
 import { getRunAgentData } from "@app/types/assistant/agent_run";
 
 const MIN_GENERATION_TOKENS = 1024;
 
 export async function ensureConversationTitle(
   authType: AuthenticatorType,
-  runAgentArgs: RunAgentArgs
+  agentLoopArgs: AgentLoopArgs
 ): Promise<string | null> {
-  const runAgentDataRes = await getRunAgentData(authType, runAgentArgs);
+  const runAgentDataRes = await getRunAgentData(authType, agentLoopArgs);
   if (runAgentDataRes.isErr()) {
     throw runAgentDataRes.error;
   }
