+++ front/temporal/agent_loop/activities/ensure_conversation_title.ts
@@ -1,10 +1,10 @@
 import { ensureConversationTitle } from "@app/lib/api/assistant/conversation/title";
 import type { AuthenticatorType } from "@app/lib/auth";
-import type { RunAgentArgs } from "@app/types/assistant/agent_run";
+import type { AgentLoopArgs } from "@app/types/assistant/agent_run";
 
 export async function ensureConversationTitleActivity(
   authType: AuthenticatorType,
-  runAgentArgs: RunAgentArgs
+  agentLoopArgs: AgentLoopArgs
 ): Promise<void> {
-  await ensureConversationTitle(authType, runAgentArgs);
+  await ensureConversationTitle(authType, agentLoopArgs);
 }
