+++ front/temporal/agent_loop/client.ts
@@ -1,53 +1,70 @@
 import { WorkflowExecutionAlreadyStartedError } from "@temporalio/client";
+import assert from "assert";
 
-import type { AuthenticatorType } from "@app/lib/auth";
+import type { Authenticator } from "@app/lib/auth";
 import { DustError } from "@app/lib/error";
+import { ConversationResource } from "@app/lib/resources/conversation_resource";
 import { getTemporalClientForAgentNamespace } from "@app/lib/temporal";
 import logger from "@app/logger/logger";
+import { logAgentLoopStart } from "@app/temporal/agent_loop/activities/instrumentation";
 import { makeAgentLoopWorkflowId } from "@app/temporal/agent_loop/lib/workflow_ids";
-import assert from "assert";
 import type { Result } from "@app/types";
 import { Err, Ok } from "@app/types";
-import type { RunAgentAsynchronousArgs } from "@app/types/assistant/agent_run";
+import type { AgentLoopArgs } from "@app/types/assistant/agent_run";
 
 import { QUEUE_NAME } from "./config";
 import { agentLoopWorkflow } from "./workflows";
 
 export async function launchAgentLoopWorkflow({
-  authType,
-  runAsynchronousAgentArgs,
+  auth,
+  agentLoopArgs,
   startStep,
-  initialStartTime,
 }: {
-  authType: AuthenticatorType;
-  runAsynchronousAgentArgs: RunAgentAsynchronousArgs;
+  auth: Authenticator;
+  agentLoopArgs: AgentLoopArgs;
   startStep: number;
   initialStartTime: number;
 }): Promise<
   Result<undefined, Error | DustError<"agent_loop_already_running">>
 > {
+  const authType = auth.toJSON();
+
+  // Capture initial start time and log total execution start.
+  const initialStartTime = Date.now();
+  const conversationId = agentLoopArgs.conversationId;
+  const agentMessageId = agentLoopArgs.agentMessageId;
+
+  logAgentLoopStart();
+  // Clear action required in conversation - the loop will put them back if needed.
+  await ConversationResource.clearActionRequired(auth, conversationId);
+
   const client = await getTemporalClientForAgentNamespace();
 
   assert(authType.workspaceId, "Workspace ID is required");
   const workflowId = makeAgentLoopWorkflowId({
     workspaceId: authType.workspaceId,
-    conversationId: runAsynchronousAgentArgs.conversationId,
-    agentMessageId: runAsynchronousAgentArgs.agentMessageId,
+    conversationId,
+    agentMessageId,
   });
 
   try {
     await client.workflow.start(agentLoopWorkflow, {
       args: [
-        { authType, runAsynchronousAgentArgs, startStep, initialStartTime },
+        {
+          authType,
+          agentLoopArgs,
+          startStep,
+          initialStartTime,
+        },
       ],
       taskQueue: QUEUE_NAME,
       workflowId,
       searchAttributes: {
-        conversationId: [runAsynchronousAgentArgs.conversationId],
+        conversationId: [conversationId],
         workspaceId: authType.workspaceId ? [authType.workspaceId] : undefined,
       },
       memo: {
-        conversationId: runAsynchronousAgentArgs.conversationId,
+        conversationId,
         workspaceId: authType.workspaceId,
       },
     });
