+++ front/temporal/agent_loop/workflows.ts
@@ -20,12 +20,12 @@ import type * as publishDeferredEventsActivities from "@app/temporal/agent_loop/
 import type * as runModelAndCreateWrapperActivities from "@app/temporal/agent_loop/activities/run_model_and_create_actions_wrapper";
 import type * as runToolActivities from "@app/temporal/agent_loop/activities/run_tool";
 import type { AgentLoopActivities } from "@app/temporal/agent_loop/lib/activity_interface";
-import { executeAgentLoop } from "@app/temporal/agent_loop/lib/agent_loop_executor";
 import { makeAgentLoopConversationTitleWorkflowId } from "@app/temporal/agent_loop/lib/workflow_ids";
 import { cancelAgentLoopSignal } from "@app/temporal/agent_loop/signals";
+import { MAX_STEPS_USE_PER_RUN_LIMIT } from "@app/types";
 import type {
-  RunAgentArgs,
-  RunAgentAsynchronousArgs,
+  AgentLoopArgs,
+  AgentLoopArgsWithTiming,
 } from "@app/types/assistant/agent_run";
 
 const toolActivityStartToCloseTimeout = `${DEFAULT_MCP_REQUEST_TIMEOUT_MS / 1000 / 60 + 1} minutes`;
@@ -86,38 +86,26 @@ const { notifyWorkflowError, finalizeCancellationActivity } = proxyActivities<
 
 export async function agentLoopConversationTitleWorkflow({
   authType,
-  runAsynchronousAgentArgs,
+  agentLoopArgs,
 }: {
   authType: AuthenticatorType;
-  runAsynchronousAgentArgs: RunAgentAsynchronousArgs;
+  agentLoopArgs: AgentLoopArgs;
 }) {
-  const runAgentArgs: RunAgentArgs = {
-    sync: false,
-    idArgs: runAsynchronousAgentArgs,
-    initialStartTime: 0,
-  };
-
-  await ensureConversationTitleActivity(authType, runAgentArgs);
+  await ensureConversationTitleActivity(authType, agentLoopArgs);
 }
 
 export async function agentLoopWorkflow({
   authType,
   initialStartTime,
-  runAsynchronousAgentArgs,
+  agentLoopArgs,
   startStep,
 }: {
   authType: AuthenticatorType;
   initialStartTime: number;
-  runAsynchronousAgentArgs: RunAgentAsynchronousArgs;
+  agentLoopArgs: AgentLoopArgs;
   startStep: number;
 }) {
   const { searchAttributes: parentSearchAttributes, memo } = workflowInfo();
-
-  const runAgentArgs: RunAgentArgs = {
-    sync: false,
-    idArgs: runAsynchronousAgentArgs,
-    initialStartTime,
-  };
 
   let childWorkflowHandle: ChildWorkflowHandle<
     typeof agentLoopConversationTitleWorkflow
@@ -137,17 +125,17 @@ export async function agentLoopWorkflow({
     // the background. If a workflow with the same ID is already running, ignore the error and
     // continue. Do not wait for the child workflow to complete at this point.
     // This is to avoid blocking the main workflow.
-    if (!runAsynchronousAgentArgs.conversationTitle) {
+    if (!agentLoopArgs.conversationTitle) {
       try {
         childWorkflowHandle = await startChild(
           agentLoopConversationTitleWorkflow,
           {
             workflowId: makeAgentLoopConversationTitleWorkflowId(
               authType,
-              runAsynchronousAgentArgs
+              agentLoopArgs
             ),
             searchAttributes: parentSearchAttributes,
-            args: [{ authType, runAsynchronousAgentArgs }],
+            args: [{ authType, agentLoopArgs }],
             memo,
           }
         );
@@ -158,10 +146,67 @@ export async function agentLoopWorkflow({
       }
     }
 
-    // In Temporal workflows, we don't pass syncStartTime since async execution doesn't need timeout.
+    const { agentMessageId, conversationId } = agentLoopArgs;
+
     await executionScope.run(async () => {
-      await executeAgentLoop(authType, runAgentArgs, activities, {
-        startStep,
+      const runIds: string[] = [];
+      const syncStartTime = Date.now();
+      let currentStep = startStep;
+
+      await activities.logAgentLoopPhaseStartActivity({
+        authType,
+        eventData: {
+          agentMessageId,
+          conversationId,
+          startStep,
+        },
+      });
+
+      for (let i = startStep; i < MAX_STEPS_USE_PER_RUN_LIMIT + 1; i++) {
+        currentStep = i;
+
+        const stepStartTime = Date.now();
+
+        const { runId, shouldContinue } = await executeStepIteration({
+          authType,
+          agentLoopArgs: {
+            ...agentLoopArgs,
+            initialStartTime,
+          },
+          activities,
+          currentStep,
+          runIds,
+          startStep,
+        });
+
+        // Update state with results.
+        if (runId) {
+          runIds.push(runId);
+        }
+
+        await activities.logAgentLoopStepCompletionActivity({
+          agentMessageId,
+          conversationId,
+          step: currentStep,
+          stepStartTime,
+        });
+
+        if (!shouldContinue) {
+          break;
+        }
+      }
+
+      const stepsCompleted = currentStep - startStep;
+
+      await activities.logAgentLoopPhaseCompletionActivity({
+        authType,
+        eventData: {
+          agentMessageId,
+          conversationId,
+          initialStartTime,
+          stepsCompleted,
+          syncStartTime,
+        },
       });
     });
 
@@ -175,12 +220,12 @@ export async function agentLoopWorkflow({
     await CancellationScope.nonCancellable(async () => {
       if (cancelRequested) {
         // Run finalization tasks on cancellation (dummy for now).
-        await finalizeCancellationActivity(authType, runAsynchronousAgentArgs);
+        await finalizeCancellationActivity(authType, agentLoopArgs);
       } else {
         await notifyWorkflowError(authType, {
-          conversationId: runAsynchronousAgentArgs.conversationId,
-          agentMessageId: runAsynchronousAgentArgs.agentMessageId,
-          agentMessageVersion: runAsynchronousAgentArgs.agentMessageVersion,
+          conversationId: agentLoopArgs.conversationId,
+          agentMessageId: agentLoopArgs.agentMessageId,
+          agentMessageVersion: agentLoopArgs.agentMessageVersion,
           error: workflowError,
         });
       }
@@ -193,4 +238,95 @@ export async function agentLoopWorkflow({
 
     throw err;
   }
+}
+
+async function executeStepIteration({
+  activities,
+  authType,
+  currentStep,
+  agentLoopArgs,
+  runIds,
+  startStep,
+}: {
+  activities: AgentLoopActivities;
+  authType: AuthenticatorType;
+  currentStep: number;
+  agentLoopArgs: AgentLoopArgsWithTiming;
+  runIds: string[];
+  startStep: number;
+}): Promise<{
+  runId: string | null;
+  shouldContinue: boolean;
+}> {
+  const result = await activities.runModelAndCreateActionsActivity({
+    authType,
+    autoRetryCount: 0,
+    checkForResume: currentStep === startStep, // Only run resume the first time.
+    runAgentArgs: agentLoopArgs,
+    runIds,
+    step: currentStep,
+  });
+
+  if (!result) {
+    // Generation completed or error occurred.
+    return {
+      runId: null,
+      shouldContinue: false,
+    };
+  }
+
+  const { runId, actionBlobs } = result;
+
+  // If at least one action needs approval, we break out of the loop and will resume once all
+  // actions have been approved.
+  const needsApproval = actionBlobs.some((a) => a.needsApproval);
+  if (needsApproval) {
+    return {
+      runId,
+      shouldContinue: false,
+    };
+  }
+
+  // Execute tools and collect any deferred events.
+  const toolResults = await Promise.all(
+    actionBlobs.map(({ actionId, retryPolicy }) =>
+      retryPolicy === "no_retry"
+        ? activities.runToolActivity(authType, {
+            actionId,
+            runAgentArgs: agentLoopArgs,
+            step: currentStep,
+            runIds: [...(runIds ?? []), ...(runId ? [runId] : [])],
+          })
+        : activities.runRetryableToolActivity(authType, {
+            actionId,
+            runAgentArgs: agentLoopArgs,
+            step: currentStep,
+            runIds: [...(runIds ?? []), ...(runId ? [runId] : [])],
+          })
+    )
+  );
+
+  // Collect all deferred events from tool executions.
+  const allDeferredEvents = toolResults.flatMap(
+    (result) => result.deferredEvents
+  );
+
+  // If there are deferred events, publish them after all tools have completed.
+  if (allDeferredEvents.length > 0) {
+    const shouldPauseWorkflow =
+      await activities.publishDeferredEventsActivity(allDeferredEvents);
+
+    if (shouldPauseWorkflow) {
+      // Break the loop - workflow will be restarted externally once required action is completed.
+      return {
+        runId,
+        shouldContinue: false,
+      };
+    }
+  }
+
+  return {
+    runId,
+    shouldContinue: !toolResults.some((result) => result.shouldPauseAgentLoop),
+  };
 }
